+++ packages/constants/index.ts
@@ -2,6 +2,7 @@ export const email = "oss@ashgw.me";
 export const repo_source = "https://github.com/ashgw/x";
 export const creator = "@rccyx";
 export const site_name = "AG";
+export const notifyEmail = "notify@ashgw.me";
 
 export const links = {
   keyBase: "https://keybase.io/ashgw",
